pyproject.toml
[project]
name = "cc-pagepulse"
version = "0.1.0"
description = "Confluence Cloud Wiki PagePulse"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "atlassian-python-api>=3.41.19",
    "beautifulsoup4>=4.13.3",
    "fastapi[all]>=0.115.8",
    "jinja2>=3.1.5",
    "openai>=1.82.1",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "toml>=0.10.2",
    "uvicorn>=0.34.0",
    "wrapt-timeout-decorator>=1.5.1",
]
